+++ api/tests/unit/application/users/index_test.js
@@ -388,25 +388,6 @@ describe('Unit | Router | user-router', function () {
     });
   });
 
-  describe('GET /api/users/{id}/authentication-methods', function () {
-    const method = 'GET';
-
-    it('returns 400 when userId is not a number', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const userId = 'wrongId';
-      const url = `/api/users/${userId}/authentication-methods`;
-
-      // when
-      const result = await httpTestServer.request(method, url);
-
-      // then
-      expect(result.statusCode).to.equal(400);
-    });
-  });
-
   describe('GET /api/users/{id}/trainings', function () {
     const method = 'GET';
